Refactor authentication code for Keycloak integration
Restructure authentication functionality by removing `Extensions.cs` and splitting its logic into `ClaimExtensions.cs` and `KeycloakServiceCollectionExtensions.cs`.

`ClaimExtensions.cs` now manages claims within a `ClaimsIdentity`, while `KeycloakServiceCollectionExtensions.cs` encapsulates the Keycloak authentication setup.

These changes improve code clarity, maintainability, and modularity for future enhancements.
